+++ Net_Core/Web.Angular/projects/gw-lib/core/account/src/c/select-preferences/select-preferences.component.ts
@@ -17,6 +17,22 @@ import { AccountService } from '../../account.service';
 import { ClientChoices, IClientChoices } from '../../client-choices.model';
 import { ISelectedableAction } from '../../selectedable-action.model';
 
+export interface IColorSchemes {
+	alternatingRow_BackColor: string;
+	back_Color: string;
+	color_Scheme: string;
+	head_Color: string;
+	header_ForeColor: string;
+	left_Color: string;
+	row_BackColor: string;
+	sub_Head_Color: string;
+}
+
+export interface IColorSchemeColumns {
+	displayedName: string;
+	propertyName: string;
+}
+
 @Component({
 	selector: 'gw-core-select-preferences',
 	standalone: true,
@@ -37,17 +53,18 @@ import { ISelectedableAction } from '../../selectedable-action.model';
 export class SelectPreferencesComponent implements OnDestroy, OnInit {
 	private _Subscription: Subscription = new Subscription();
 
+	colorSchemeColumns: IColorSchemeColumns[] = [];
 	clientChoices: IClientChoices = new ClientChoices();
 	frmSelectPreferences!: FormGroup;
 	selectedColorScheme!: string;
 	selectedAction!: string;
 
-	validColorSchemes = [
-		{ color_Scheme: 'Blue'	  ,head_Color: '#C7C7C7'	,header_ForeColor: 'Black'	,row_BackColor: '#b6cbeb'	,alternatingRow_BackColor: '#6699cc'	,sub_Head_Color: '#b6cbeb'	,back_Color: '#ffffff'	,left_Color: '#eeeeee' },
-		{ color_Scheme: 'Green'	  ,head_Color: '#808577'	,header_ForeColor: 'White'	,row_BackColor: '#879966'	,alternatingRow_BackColor: '#c5e095'	,sub_Head_Color: '#879966'	,back_Color: '#ffffff'	,left_Color: '#eeeeee' },
-		{ color_Scheme: 'Yellow'	,head_Color: '#CF9C00'	,header_ForeColor: 'Black'	,row_BackColor: '#f8bc03'	,alternatingRow_BackColor: '#f8e094'	,sub_Head_Color: '#f8bc03'	,back_Color: '#ffffff'	,left_Color: '#f8e094' },
-		{ color_Scheme: 'Purple'	,head_Color: '#C7C7C7'	,header_ForeColor: 'Black'	,row_BackColor: '#be9cc5'	,alternatingRow_BackColor: '#91619b'	,sub_Head_Color: '#be9cc5'	,back_Color: '#ffffff'	,left_Color: '#eeeeee' },
-		{ color_Scheme: 'Red'	    ,head_Color: '#BA706A'	,header_ForeColor: 'White'	,row_BackColor: '#DE8587'	,alternatingRow_BackColor: '#A72A49'	,sub_Head_Color: '#df867f'	,back_Color: '#ffffff'	,left_Color: '#eeeeee' }
+	validColorSchemes: IColorSchemes[] = [
+		{ alternatingRow_BackColor: '#6699cc'	,back_Color: '#ffffff'	,color_Scheme: 'Blue'	,head_Color: '#C7C7C7'	,header_ForeColor: 'Black'	,left_Color: '#eeeeee'	,row_BackColor: '#b6cbeb'	, sub_Head_Color: '#b6cbeb' },
+		{ alternatingRow_BackColor: '#c5e095'	,back_Color: '#ffffff'	,color_Scheme: 'Green'	,head_Color: '#808577'	,header_ForeColor: 'White'	,left_Color: '#eeeeee'	,row_BackColor: '#879966'	, sub_Head_Color: '#879966' },
+		{ alternatingRow_BackColor: '#f8e094'	,back_Color: '#ffffff'	,color_Scheme: 'Yellow'	,head_Color: '#CF9C00'	,header_ForeColor: 'Black'	,left_Color: '#f8e094'	,row_BackColor: '#f8bc03'	, sub_Head_Color: '#f8bc03' },
+		{ alternatingRow_BackColor: '#91619b'	,back_Color: '#ffffff'	,color_Scheme: 'Purple'	,head_Color: '#C7C7C7'	,header_ForeColor: 'Black'	,left_Color: '#eeeeee'	,row_BackColor: '#be9cc5'	, sub_Head_Color: '#be9cc5' },
+		{ alternatingRow_BackColor: '#A72A49'	,back_Color: '#ffffff'	,color_Scheme: 'Red'    ,head_Color: '#BA706A'	,header_ForeColor: 'White'	,left_Color: '#eeeeee'	,row_BackColor: '#DE8587'	, sub_Head_Color: '#df867f' }
 	];
 
 	validLinks = [
@@ -68,6 +85,7 @@ export class SelectPreferencesComponent implements OnDestroy, OnInit {
 	}
 
 	ngOnInit(): void {
+		this.setHeaders();
 		this.frmSelectPreferences = new FormGroup({
 			recordsPerPage: new FormControl(10),
 			selectedColorScheme: new FormControl('Blue'),
@@ -82,6 +100,13 @@ export class SelectPreferencesComponent implements OnDestroy, OnInit {
 
 	get controls() {
 		return this.frmSelectPreferences.controls;
+	}
+
+	// eslint-disable-next-line @typescript-eslint/no-explicit-any
+	getValue(colorSchemes: IColorSchemes, columnName: IColorSchemeColumns): string {
+		type ObjectKey = keyof typeof colorSchemes;
+		const mKey: ObjectKey = columnName.propertyName as ObjectKey;
+		return colorSchemes[mKey];
 	}
   
 	onSubmit(): void {
@@ -123,5 +148,19 @@ export class SelectPreferencesComponent implements OnDestroy, OnInit {
 				selectedColorScheme: ['Blue', [Validators.required]],
 			});
 		}
+	}
+
+	setHeaders(): void {
+		// loop through IColorSchemes
+		for (const [key] of Object.entries(this.validColorSchemes[0])) {
+			const mWords = key.split('_');
+			const mDisplayedName = mWords.map((w) => w.charAt(0).toUpperCase() + w.slice(1)).join(' ');
+			const mColorSchemeColumn = {
+				displayedName: mDisplayedName,
+				propertyName: key
+			};
+			this.colorSchemeColumns.push(mColorSchemeColumn);
+		  }
+		//   console.log('xx.setHeaders', this.colorSchemeColumns);
 	}
 }