+++ Release_Notes.md
@@ -8,6 +8,7 @@
 - Connection strings are now handled in DDatabaseManager
 - AbstractDBInteraction.BulkInsert now uses a DTO for the parameter making it generic
 - Vertical menu items now hightlight when clicked on
+- Select Preferences Component html builds from typescript data
 
 ### Bug Fixes
 - Fixed type-o in Change Password